# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibewatt",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/wrapped": {
      "get": {
        "summary": "Wrapped",
        "operationId": "wrapped_api_wrapped_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 9998,
                  "minimum": 1970
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "claude-code",
                "cowork",
                "web"
              ],
              "type": "string",
              "default": "all",
              "title": "Source"
            }
          },
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "cost",
                "tokens"
              ],
              "type": "string",
              "default": "cost",
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/concierge": {
      "get": {
        "summary": "Concierge",
        "operationId": "concierge_api_concierge_get",
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciergeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/findings": {
      "get": {
        "summary": "Findings",
        "description": "The stored snapshot; recomputed only after the store changes (A-092).",
        "operationId": "findings_api_findings_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "anomaly",
                    "cache",
                    "tip",
                    "waste",
                    "peak",
                    "context"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "info",
                    "warning",
                    "urgent"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "include_dismissed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Dismissed"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "claude-code",
                "cowork",
                "web"
              ],
              "type": "string",
              "default": "all",
              "title": "Source"
            }
          },
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "cost",
                "tokens"
              ],
              "type": "string",
              "default": "cost",
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/findings/{finding_id}": {
      "get": {
        "summary": "Finding Detail",
        "description": "One finding, for the deep-linkable finding modal (A-046).",
        "operationId": "finding_detail_api_findings__finding_id__get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingOut"
                }
              }
            }
          },
          "404": {
            "description": "No such finding"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/findings/{finding_id}/dismiss": {
      "post": {
        "summary": "Dismiss Finding",
        "operationId": "dismiss_finding_api_findings__finding_id__dismiss_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissFindingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DismissFindingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{session_id}": {
      "get": {
        "summary": "Session Detail",
        "operationId": "session_detail_api_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sync": {
      "post": {
        "summary": "Sync",
        "operationId": "sync_api_sync_post",
        "responses": {
          "200": {
            "description": "With `Accept: application/x-ndjson`, one `{done, total}` line per batch, then the result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultOut"
                }
              },
              "application/x-ndjson": {}
            }
          },
          "409": {
            "description": "A sync is already running"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOut": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "type": "array",
            "title": "Findings"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "anomaly_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anomaly Notes",
            "default": []
          },
          "analyzed_at": {
            "type": "string",
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "findings",
          "notes",
          "analyzed_at"
        ],
        "title": "AnalysisOut"
      },
...
      "ConciergeOut": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "project",
          "text",
          "notes"
        ],
        "title": "ConciergeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}